Linux compatibility for examples and ctrl c escape issue (#16)
* Better Linux compatibility feature update on examples

Added additional feature flags for using eframes in x11 and wayland environments.
When compiling, use `cargo run --features wayland` in wayland Ubuntu environment.

* Ctrl+C (Esacpe) is not working in Linux (or Windows)

This patch  has not been tested on Windows, but in most environments except macOS, Ctrl+C is recognized as a Copy event via egui::Event. This makes it difficult to trigger the important Escape Event in the terminal. This patch temporarily makes it work as a Copy Event when characters on the screen are selected, and as an Escape when they are not selected, but not on macOS.
In future, the method bypassing egui's normal event interpreter would be better.

* Update src/view.rs

Better approach instead judged by selected mode

Co-authored-by: Ilia Shvyrialkin <ilyashvy@gmail.com>

* Ctrl+P additional addressing comment

---------

Co-authored-by: Ilia Shvyrialkin <ilyashvy@gmail.com>